requirements.txt
-i https://pypi.org/simple
black==25.1.0; python_version >= '3.9'
certifi==2025.6.15; python_version >= '3.7'
cfgv==3.4.0; python_version >= '3.8'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
colorama==0.4.6
contourpy==1.3.2; python_version >= '3.10'
cycler==0.12.1; python_version >= '3.8'
distlib==0.3.9
et-xmlfile==2.0.0; python_version >= '3.8'
filelock==3.18.0; python_version >= '3.9'
flake8==7.3.0; python_version >= '3.9'
fonttools==4.58.4; python_version >= '3.9'
fsspec==2025.5.1; python_version >= '3.9'
hf-xet==1.1.5; platform_machine == 'x86_64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'aarch64'
huggingface-hub==0.33.1; python_full_version >= '3.8.0'
identify==2.6.12; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
ijson==3.4.0; python_version >= '3.9'
iniconfig==2.1.0; python_version >= '3.8'
isort[colors]==6.0.1; python_full_version >= '3.9.0'
jinja2==3.1.6; python_version >= '3.7'
joblib==1.5.1; python_version >= '3.9'
kiwisolver==1.4.8; python_version >= '3.10'
markupsafe==3.0.2; python_version >= '3.9'
matplotlib==3.10.3; python_version >= '3.10'
mccabe==0.7.0; python_version >= '3.6'
mpmath==1.3.0
mypy-extensions==1.1.0; python_version >= '3.8'
networkx==3.5; python_version >= '3.11'
nodeenv==1.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
numpy==2.3.1; python_version >= '3.11'
openpyxl==3.1.5; python_version >= '3.8'
packaging==25.0; python_version >= '3.8'
pandas==2.3.0; python_version >= '3.9'
pathspec==0.12.1; python_version >= '3.8'
pillow==11.2.1; python_version >= '3.9'
platformdirs==4.3.8; python_version >= '3.9'
pluggy==1.6.0; python_version >= '3.9'
polars==1.31.0; python_version >= '3.9'
pre-commit==4.2.0; python_version >= '3.9'
pycodestyle==2.14.0; python_version >= '3.9'
pyflakes==3.4.0; python_version >= '3.9'
pygments==2.19.2; python_version >= '3.8'
pyparsing==3.2.3; python_version >= '3.9'
pytest==8.4.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.4; python_version >= '3.8'
safetensors==0.5.3; python_version >= '3.7'
scikit-learn==1.7.0; python_version >= '3.10'
scipy==1.16.0; python_version >= '3.11'
seaborn==0.13.2; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sympy==1.14.0; python_version >= '3.9'
threadpoolctl==3.6.0; python_version >= '3.9'
tokenizers==0.21.2; python_version >= '3.9'
torch==2.7.1; python_full_version >= '3.9.0'
tqdm==4.67.1; python_version >= '3.7'
transformers==4.53.0; python_full_version >= '3.9.0'
typing-extensions==4.14.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
urllib3==2.5.0; python_version >= '3.9'
virtualenv==20.31.2; python_version >= '3.8'
